+++ xc3_model/src/lib.rs
@@ -41,7 +41,6 @@ use indexmap::IndexMap;
 use log::error;
 use material::create_materials;
 use shader_database::ShaderDatabase;
-use smol_str::SmolStr;
 use texture::{load_textures, load_textures_legacy};
 use thiserror::Error;
 use vertex::ModelBuffers;
@@ -548,11 +547,11 @@ pub fn load_model<P: AsRef<Path>>(
     let streaming_data = StreamingData::new(&mxmd, &wismt_path, is_pc, chr_tex_folder.as_deref())?;
 
     let model_name = model_name(wimdo_path);
-    let spch = shader_database.and_then(|database| database.files.get(&SmolStr::from(&model_name)));
+    let model_programs = shader_database.and_then(|database| database.model(&model_name));
 
     let chr = load_chr(wimdo_path, model_name);
 
-    ModelRoot::from_mxmd_model(&mxmd, chr, &streaming_data, spch)
+    ModelRoot::from_mxmd_model(&mxmd, chr, &streaming_data, model_programs.as_ref())
 }
 
 fn load_chr(wimdo_path: &Path, model_name: String) -> Option<Sar1> {