+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 * Optimized the performance for loading the shader database JSON file for xc3_model.
 * Adjusted decompiled shader annotation to replace constant buffer parameters with their values.
+* Changed `xc3_model::shader_database::ShaderDatabase` to be an opaque wrapper. Get data using the `model` and `map` methods.
 
 ### Fixed
 * Fixed an issue where some Xenoblade 3 shader database entires assigned normal maps to glossiness.